requirements.txt
tensorflow
sklearn
tqdm
pandas
biopython
pillow
matplotlib
shap
snakemake
pyyaml
opencv-python